Revert "feature/#1/implement-bot-class"
