+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "start": "node index.js",
+    "start": "node src/discoji.js",
     "test": "node test/tests.js"
   },
   "author": "marc",
