+++ src/command.class.js
@@ -41,8 +41,7 @@ class Command {
   }
 
   isValid() {
-    if (this.userId !== this.prefix || !this.tokens.length)
-      return new Error('Your message has a wrong format');
+    return this.userId === this.prefix && this.tokens.length;
   }
 
   structure() {
@@ -66,7 +65,8 @@ class Command {
     }
 
     if (this.backgroundIcon === this.fontIcon)
-      return new Error('No way, you are going nowhere with this combination of emojis');
+      throw new Error(
+        'No way, you are going nowhere with this combination of emojis');
   }
 
   set message(value) {
