+++ src/emojiencoder.class.js
